Split middleware test suite sources
