+++ triggers/service/auth/src/test/scala/com/daml/auth/middleware/oauth2/TestMiddleware.scala
@@ -8,7 +8,7 @@ import java.time.Duration
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.headers.{Cookie, Location, `Set-Cookie`}
-import com.daml.auth.middleware.api.{Client, Request, RequestStore}
+import com.daml.auth.middleware.api.{Client, Request}
 import com.daml.auth.middleware.api.Tagged.{AccessToken, RefreshToken}
 import com.daml.jwt.JwtSigner
 import com.daml.jwt.domain.DecodedJwt
@@ -20,9 +20,8 @@ import com.daml.auth.oauth2.api.{Response => OAuthResponse}
 import org.scalatest.wordspec.AsyncWordSpec
 
 import scala.util.{Failure, Success}
-import scala.concurrent.duration._
 
-class Test extends AsyncWordSpec with TestFixture with SuiteResourceManagementAroundAll {
+class TestMiddleware extends AsyncWordSpec with TestFixture with SuiteResourceManagementAroundAll {
   private def makeToken(
       claims: Request.Claims,
       secret: String = "secret",
@@ -258,7 +257,7 @@ class Test extends AsyncWordSpec with TestFixture with SuiteResourceManagementAr
   }
 }
 
-class TestCallbackUriOverride
+class TestMiddlewareCallbackUriOverride
     extends AsyncWordSpec
     with TestFixture
     with SuiteResourceManagementAroundAll {
@@ -284,7 +283,7 @@ class TestCallbackUriOverride
   }
 }
 
-class TestLimitedMiddlewareCallbackStore
+class TestMiddlewareLimitedCallbackStore
     extends AsyncWordSpec
     with TestFixture
     with SuiteResourceManagementAroundAll {
@@ -331,7 +330,7 @@ class TestLimitedMiddlewareCallbackStore
   }
 }
 
-class TestLimitedClientCallbackStore
+class TestMiddlewareClientLimitedCallbackStore
     extends AsyncWordSpec
     with TestFixture
     with SuiteResourceManagementAroundAll {
@@ -385,63 +384,5 @@ class TestLimitedClientCallbackStore
         assert(resultCarol.status == StatusCodes.OK)
       }
     }
-  }
-}
-
-class TestRequestStore extends AsyncWordSpec {
-  "return None on missing element" in {
-    val store = new RequestStore[Int, String](1, 1.day)
-    assert(store.pop(0) == None)
-  }
-
-  "return previously put element" in {
-    val store = new RequestStore[Int, String](1, 1.day)
-    store.put(0, "zero")
-    assert(store.pop(0) == Some("zero"))
-  }
-
-  "return None on previously popped element" in {
-    val store = new RequestStore[Int, String](1, 1.day)
-    store.put(0, "zero")
-    store.pop(0)
-    assert(store.pop(0) == None)
-  }
-
-  "store multiple elements" in {
-    val store = new RequestStore[Int, String](3, 1.day)
-    store.put(0, "zero")
-    store.put(1, "one")
-    store.put(2, "two")
-    assert(store.pop(0) == Some("zero"))
-    assert(store.pop(1) == Some("one"))
-    assert(store.pop(2) == Some("two"))
-  }
-
-  "store no more than max capacity" in {
-    val store = new RequestStore[Int, String](2, 1.day)
-    assert(store.put(0, "zero"))
-    assert(store.put(1, "one"))
-    assert(!store.put(2, "two"))
-    assert(store.pop(0) == Some("zero"))
-    assert(store.pop(1) == Some("one"))
-    assert(store.pop(2) == None)
-  }
-
-  "return None on timed out element" in {
-    var time: Long = 0
-    val store = new RequestStore[Int, String](1, 1.day, () => time)
-    store.put(0, "zero")
-    time += 1.day.toNanos
-    assert(store.pop(0) == None)
-  }
-
-  "free capacity for timed out elements" in {
-    var time: Long = 0
-    val store = new RequestStore[Int, String](1, 1.day, () => time)
-    assert(store.put(0, "zero"))
-    assert(!store.put(1, "one"))
-    time += 1.day.toNanos
-    assert(store.put(2, "two"))
-    assert(store.pop(2) == Some("two"))
   }
 }
